Refactor implementation into correct builders
This bugged me for a while, so fixing it now before addressing more
issues in the hierarchy builder
